requirements.txt
Flask==2.3.3
pandas==2.1.0
Werkzeug==2.3.7
numpy==1.25.2
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
gunicorn==21.2.0
flask-cors==4.0.0
mysql-connector-python==8.1.0
SQLAlchemy==2.0.23 